server/requirements.txt
appdirs==1.4.3
certifi==2019.11.28
chardet==3.0.4
distlib==0.3.0
Django==3.0.5
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-js-asset==1.2.2
django-taggit==1.2.0
djangorestframework==3.9.4
Pillow==6.0.0
python-dateutil==2.8.1
pytz==2019.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
stripe==2.44.0
urllib3==1.25.8
